Fix error in `minimum`, `maximum`
Both methods were automatically skipping null
entries, regardless of the value of `skipnull`,
for `X::NullableArray` arguments of length > 16.

This behavior was due to the respective 
`Base.mapreduce_impl` methods, which I had 
implemented before the `_mapreduce_skipnull`
interface.

The previous tests did not catch this behavior
because they did not test mapreduce-related
methods on argument `NullableArray`s of length >
16. These tests, and tests for `src/statistics.jl`
are updated accordingly.