Update show methods to Julia 0.5
Since JuliaLang/julia#15928, Nullable() is printed as #NULL, which
means we no longer need to override show methods from Base.
This fixes breakage in current Julia 0.5.
Also fix a bug in show(), where NullableArray{T,N} was printed as the
element type prefix: add parentheses to make it clear this is the type
of the container.